Breakout PolygonDrawable
